requirements.txt
playwright
pytest-playwright
pytest
allure-pytest
pytest-dotenv
python-slugify